Don't clear icon cache on manual refresh
and increase HTTP cache size.

Fixes #1230

+++ mobile/src/main/java/org/openhab/habdroid/util/CacheManager.kt
@@ -16,11 +16,9 @@ package org.openhab.habdroid.util
 import android.content.Context
 import android.graphics.Bitmap
 import android.util.LruCache
-
 import okhttp3.Cache
 import okhttp3.HttpUrl
 import org.openhab.habdroid.model.IconFormat
-
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
@@ -28,8 +26,9 @@ import java.io.IOException
 import java.io.InputStream
 
 class CacheManager private constructor(appContext: Context) {
-    val httpCache: Cache = Cache(File(appContext.cacheDir, "http"), (2 * 1024 * 1024).toLong())
-    private val bitmapCache: LruCache<HttpUrl, Bitmap>
+    val httpCache: Cache = Cache(File(appContext.cacheDir, "http"), (10 * 1024 * 1024).toLong())
+    private val iconBitmapCache: LruCache<HttpUrl, Bitmap>
+    private val temporaryBitmapCache: LruCache<HttpUrl, Bitmap>
     private val widgetIconDirectory = appContext.getDir("widgeticons", Context.MODE_PRIVATE)
 
     init {
@@ -38,7 +37,12 @@ class CacheManager private constructor(appContext: Context) {
         // int in its constructor.
         val maxMemory = (Runtime.getRuntime().maxMemory() / 1024).toInt()
         // Use up to 1/8 of the available VM memory for the bitmap cache
-        bitmapCache = object : LruCache<HttpUrl, Bitmap>(maxMemory / 8) {
+        iconBitmapCache = object : LruCache<HttpUrl, Bitmap>(maxMemory / 8) {
+            override fun sizeOf(key: HttpUrl, value: Bitmap): Int {
+                return value.byteCount / 1024
+            }
+        }
+        temporaryBitmapCache = object : LruCache<HttpUrl, Bitmap>(maxMemory / 8) {
             override fun sizeOf(key: HttpUrl, value: Bitmap): Int {
                 return value.byteCount / 1024
             }
@@ -46,11 +50,24 @@ class CacheManager private constructor(appContext: Context) {
     }
 
     fun getCachedBitmap(url: HttpUrl): Bitmap? {
-        return bitmapCache.get(url)
+        return if (isIconUrl(url)) {
+            iconBitmapCache.get(url)
+        } else {
+            temporaryBitmapCache.get(url)
+        }
     }
 
     fun cacheBitmap(url: HttpUrl, bitmap: Bitmap) {
-        bitmapCache.put(url, bitmap)
+        if (isIconUrl(url)) {
+            iconBitmapCache.put(url, bitmap)
+        } else {
+            temporaryBitmapCache.put(url, bitmap)
+        }
+    }
+
+    private fun isIconUrl(url: HttpUrl): Boolean {
+        return url.pathSegments.firstOrNull() == "icon" &&
+            url.pathSegments[1].isNotEmpty()
     }
 
     fun isBitmapCached(url: HttpUrl): Boolean {
@@ -69,8 +86,7 @@ class CacheManager private constructor(appContext: Context) {
 
     fun getWidgetIconFormat(widgetId: Int): IconFormat? {
         return IconFormat.values()
-            .filter { format -> getWidgetIconFile(widgetId, format).exists() }
-            .firstOrNull()
+            .firstOrNull { format -> getWidgetIconFile(widgetId, format).exists() }
     }
 
     fun getWidgetIconStream(widgetId: Int): InputStream? {
@@ -79,14 +95,17 @@ class CacheManager private constructor(appContext: Context) {
         }
     }
 
-    fun clearCache() {
-        bitmapCache.evictAll()
+    fun clearCache(alsoClearIcons: Boolean) {
+        temporaryBitmapCache.evictAll()
         try {
             httpCache.evictAll()
         } catch (ignored: IOException) {
             // ignored
         }
-        widgetIconDirectory?.listFiles()?.forEach { f -> f.delete() }
+        if (alsoClearIcons) {
+            widgetIconDirectory?.listFiles()?.forEach { f -> f.delete() }
+            iconBitmapCache.evictAll()
+        }
     }
 
     private fun getWidgetIconFile(widgetId: Int, format: IconFormat): File {
